requirements.txt
openpyxl==3.0.9
Pillow==9.1.0